refactoring to use es6 modules
