+++ package.json
@@ -27,17 +27,19 @@
     "uuid": "2.0.1"
   },
   "devDependencies": {
-    "babel": "5.8.23",
-    "babel-core": "5.8.24",
-    "babel-loader": "5.3.2",
+    "babel": "^6.5.2",
+    "babel-core": "^6.8.0",
+    "babel-loader": "^6.2.4",
+    "babel-preset-react-native-stage-0": "^1.0.1",
+    "babel-register": "^6.8.0",
     "babelify": "7.2.0",
     "chai": "3.3.0",
     "commitizen": "1.0.4",
     "cz-conventional-changelog": "1.1.2",
     "eslint": "^2.8.0",
     "eslint-loader": "^1.3.0",
-    "mocha": "2.3.3",
-    "mocha-phantomjs": "4.0.1",
+    "mocha": "^2.4.5",
+    "mocha-phantomjs": "^4.0.2",
     "require": "2.4.20",
     "rimraf": "2.5.0",
     "webpack": "1.12.1"
