+++ src/index.js
@@ -1,19 +1,20 @@
 'use strict';
-var Commands = require('./commands');
-var filter = require('./filter');
-var SCSpatialFeature = require('./sc.spatialfeature.js');
-var SCGeometry = require('./sc.geometry');
-var SC = require('./sc');
+
+import Commands from './commands';
+import Filter from './filter';
+import SCSpatialFeature from './sc.spatialfeature';
+import SCGeometry from './sc.geometry';
+import SC from './core';
 
 var SpatialConnect = (function() {
   var sc = {};
   sc.action = SC.action;
   sc.stream = SC.stream;
-  sc.Filter = filter;
+  sc.filter = Filter;
   sc.gfeature = SCGeometry;
   sc.feature = SCSpatialFeature;
   sc.Commands = Commands;
   return sc;
 })();
 
-module.exports = SpatialConnect;
+export default SpatialConnect;
