+++ test/index.html
@@ -3,7 +3,6 @@
     <head>
         <title>Simple Tests - Using Mocha & Chai</title>
         <link rel="stylesheet" href="../node_modules/mocha/mocha.css" />
-        <script src="../node_modules/openlayers/dist/ol-debug.js"></script>
     </head>
     <body>
         <div id="mocha"></div> <!-- 1 -->
